+++ addon/initializers/offline-globals.js
@@ -2,19 +2,17 @@ export function initialize(application) {
   //Inject OfflieGlobalsService onto factories of specified types.
   [
     'store:local',
-    'syncer:main',
-    'component',
     'controller',
+    'component',
     'model',
     'route',
-    'view'
-  ].forEach(type => {
+    'view',
+  ].forEach((type) => {
     application.inject(type, 'offlineGlobals', 'service:offline-globals');
   });
 }
 
 export default {
   name: 'offline-globals',
-  after: 'syncer',
-  initialize
+  initialize,
 };